refactor: separate Medical from TableBody
